requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# OCR & Document Processing
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==10.1.0
camelot-py[cv]==0.11.0
pdfplumber==0.10.3
opencv-python==4.8.1.78

# LLM & AI
langchain==0.1.0
langchain-openai==0.0.2
openai==1.6.1
tiktoken==0.5.2

# Embeddings & Vector DB
sentence-transformers==2.2.2
chromadb==0.4.22

# ML & Anomaly Detection
scikit-learn==1.3.2
prophet==1.1.5
statsmodels==0.14.1
numpy==1.24.3
pandas==2.1.4

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Storage
boto3==1.34.10

# Utilities
python-multipart==0.0.6
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1

# Monitoring & Logging
prometheus-client==0.19.0
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Security
cryptography==41.0.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
